feat: 🎸 add useFocusable
